fix(inventario): estandariza stock como entero y mejora manejo de precios
- Integra autoload PSR-4 en index.php (corrige "Class not found")
- Fuerza uso de int en stock en servicios y rutas
- Actualiza DB: stock → INT UNSIGNED, costos → DECIMAL(12,2)
- UI: stock sin decimales (step=1) y redondeo en vistas
- Añade campo precio de venta y máscara con separadores de miles
- Corrige error de sintaxis en ProductsPage.js (preview de imágenes)

BREAKING CHANGE: cambio de tipo en campos de stock en base de datos

+++ src/API/routes/inventario.php
@@ -298,7 +298,7 @@ function handleGetResumenInventario(): void {
             'dias' => $dias,
             'resumen' => [
                 'total_alertas' => (int)($row['total_alertas'] ?? 0),
-                'stock_en_riesgo' => (float)($row['stock_en_riesgo'] ?? 0),
+                'stock_en_riesgo' => (int)($row['stock_en_riesgo'] ?? 0),
                 'vencidos' => (int)($row['vencidos'] ?? 0),
                 'rojos' => (int)($row['rojos'] ?? 0),
                 'amarillos' => (int)($row['amarillos'] ?? 0),
@@ -467,8 +467,9 @@ function handlePostMovimientoInventario(): void {
             'movimiento_id' => $movId,
             'lote' => [
                 'id' => (int)$lote['id'],
-                'stock_actual' => (float)$lote['stock_actual'],
-                'stock_reservado' => (float)$lote['stock_reservado'],
+                'stock_actual' => (int)$lote['stock_actual'],
+                'stock_reservado' => (int)$lote['stock_reservado'],
+
             ],
         ], 201);
     } catch (Throwable $e) {
